Add initial version of TweetDeleter
This commit as the initial implementation of TweetDeleter.
This works fine as-is. It requires chromedp to run in a non-headless
mode to successfully navigate the Javascript to delete. Program
will stop executing once all tweets are removed, but there's a small
bug where program will hang if initial time range contains no tweets
to delete.